refactor: rename NPM_GLOBAL_PREFIX env var to QD_NPM_PREFIX
