+++ quarkdown-html/src/test/kotlin/com/quarkdown/rendering/html/HtmlToPdfTest.kt
@@ -1,6 +1,6 @@
 package com.quarkdown.rendering.html
 
-import com.quarkdown.interaction.executable.Env
+import com.quarkdown.interaction.Env
 import com.quarkdown.interaction.executable.NodeJsWrapper
 import com.quarkdown.interaction.executable.NpmWrapper
 import com.quarkdown.rendering.html.pdf.HtmlPdfExportOptions
@@ -40,7 +40,7 @@ class HtmlToPdfTest {
 
     @Test
     fun `bare script on simple html`() {
-        assumeTrue(Env.npmGlobalPrefix != null)
+        assumeTrue(Env.npmPrefix != null)
         assumeTrue(Env.nodePath != null)
         val node = NodeJsWrapper(options.nodeJsPath, workingDirectory = directory)
         assumeTrue(node.isValid)
