+++ quarkdown-interaction/src/main/kotlin/com/quarkdown/interaction/executable/NodeModule.kt
@@ -1,5 +1,7 @@
 package com.quarkdown.interaction.executable
 
+import com.quarkdown.interaction.Env
+
 /**
  * Abstraction of a Node.js module, which can be installed and linked to a [NodeJsWrapper] through a [NpmWrapper].
  * @param name name of the module, which matches the name in the NPM registry
@@ -16,8 +18,8 @@ class NodeModuleNotInstalledException(
     module: NodeModule,
 ) : IllegalStateException(
         """
-        Module '${module.name}' is not installed. Please install it via `npm install ${module.name} --prefix $${Env.npmGlobalPrefix}` and retry.
-        Make sure ${Env.npmGlobalPrefix} is an environment variable pointing to Quarkdown's `lib` directory or any other installation directory,
+        Module '${module.name}' is not installed. Please install it via `npm install ${module.name} --prefix $${Env.npmPrefix}` and retry.
+        Make sure ${Env.npmPrefix} is an environment variable pointing to Quarkdown's `lib` directory or any other installation directory,
         and it must be available at Quarkdown's launch.
         
         For more information, see: https://github.com/iamgio/quarkdown/wiki/pdf-export
